==> code/06_PAV_soybean_pangenome.Rmd <==
---
title: "Candidate gene PAV in the soybean pangenome"
author: "Fabricio Almeida-Silva"
date: "5/26/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=TRUE,
                      warning=FALSE, 
                      message=FALSE,
                      tidy.opts=list(width.cutoff=60),
                      tidy=TRUE)
```

## PAV in the soybean pangenome

Here, we will analyze patterns of presence/absence variation (PAV) for all mined candidate genes in the soybean pangenome. The data was retrieved from Torkamaneh *et al.,* 2021.

```{r PAV}
library(here)
library(tidyverse)
options(timeout=300)
pan_file <- here("data", "pansoy.xlsx")
if(!file.exists(pan_file)) {
  download.file("https://soybase.org/data/public/Glycine_max/Wm82.gnm4.pan.W46N/glyma.Wm82.gnm4.pan.W46N.PanSoy_supp_tables.xlsx", destfile = pan_file)
  
}

# Load PAV matrix
pav <- readxl::read_xlsx(here("data", "pansoy.xlsx"), sheet=4, skip=2)
pav <- as.data.frame(pav)

# Create tables of candidates
load(here("products", "result_files", "aglycines_minedcand.rda"))
load(here("products", "result_files", "slitura_minedcand.rda"))
load(here("products", "result_files", "hglycines_minedcand.rda"))


# Count frequency of genes in the genomes
pav_agl <- pav[pav$`Gene/ Accession` %in% mining_aglycines$gene, c(1,207)]
pav_sli <- pav[pav$`Gene/ Accession` %in% mining_slitura$gene, c(1,207)]
pav_hgl <- pav[pav$`Gene/ Accession` %in% mining_hglycines$gene, c(1,207)]
freq_table <- rbind(pav_agl, pav_sli, pav_hgl)
freq_table$Species <- c(
    rep("Aglycines", nrow(pav_agl)),
    rep("Slitura", nrow(pav_sli)),
    rep("Hglycines", nrow(pav_hgl))
)
save(freq_table, 
     file = here("products", "result_files", "pav_pangenome_table.rda"),
     compress="xz")

# Fisher's exact test to see if the dispensable genome is enriched in candidates
cand_dispensable <- unique(freq_table[freq_table$Mean < 0.99, ][,1])
cand_notdispensable <- unique(freq_table[,1][!freq_table[,1] %in% cand_dispensable])
dispensable <- pav[,1][pav$Mean < 0.99]
dispensable_notcand <- dispensable[!dispensable %in% freq_table[,1]]
notdispensable <- pav[,1][pav$Mean >= 0.99]
notdispensable_notcand <- notdispensable[!notdispensable %in% freq_table[,1]]
cont_table <- matrix(
    c(length(cand_dispensable), length(cand_notdispensable),
      length(dispensable_notcand), length(notdispensable_notcand)),
    byrow = TRUE, nrow=2, ncol=2
)
fisher.test(cont_table, alternative = "greater")

# Plot points
library(ggsci)
pav_plot <- freq_table %>%
    arrange(Species, -Mean) %>%
    mutate(index=1:nrow(freq_table)) %>%
    mutate(key = case_when(Mean < 0.98 ~ `Gene/ Accession`,
                           TRUE ~ "")) %>%
    ggplot(., aes(x=index, y=Mean, label=key)) +
    geom_point(aes(color=Species), size=1) +
    scale_color_aaas() +
    ggrepel::geom_text_repel(size = 3) +
    theme_classic() +
    labs(title="PAV of mined candidate genes in the soybean pangenome",
         x="Gene index", y="Relative frequency")

ggsave(pav_plot, filename = here("products", "plots", "pav.png"),
       dpi = 300)    
```

Now, let's visualize it as a heatmap of presence/absence.

```{r binary_heatmap}
# Create matrix
mat <- pav[pav$`Gene/ Accession` %in% 
               c(mining_aglycines$gene, 
                 mining_slitura$gene,
                 mining_hglycines$gene), ]
mat <- as.data.frame(mat)
rownames(mat) <- mat$`Gene/ Accession`
mat$`Gene/ Accession` <- NULL
mat <- mat[, 1:205]

# Load accession annotation
coldata <- readxl::read_xlsx(here("data", "pansoy.xlsx"), sheet=1, skip=2)
coldata <- as.data.frame(coldata)
rownames(coldata) <- coldata$`GmHapMap ID`
coldata <- coldata[, 3, drop=FALSE]
colnames(coldata) <- "Origin"

# Reorder columns and rows to match categories and remove low frequency countries
count <- table(coldata$Origin)
other <- names(count[count < 2])
col_annot <- coldata
col_annot$Origin[col_annot$Origin %in% other] <- "Other"
col_annot <- col_annot[order(col_annot$Origin), , drop=FALSE]
matplot <- mat[, rownames(col_annot)]
matplot <- ceiling(matplot)
n <- length(unique(col_annot$Origin))

# Create custom annotation colors
ann_colors <- list(Origin = ggsci::pal_d3("category20")(n))
names(ann_colors$Origin) <- unique(col_annot$Origin)

grid::gpar(fontface = "plain")
library(pheatmap)
hm2 <- pheatmap::pheatmap(as.matrix(matplot), 
                          color = c("darkseagreen1", "springgreen4"),
                          annotation_col = col_annot, 
                          annotation_colors = ann_colors,
                          cluster_cols = FALSE,
                          show_rownames = FALSE, 
                          fontsize_col = 5,
                          legend=FALSE,
                          border_color = NA)

hm_ggplot <- ggplotify::as.ggplot(hm2)
hm_ggplot <- hm_ggplot + 
    ggtitle("PAV per accession and their geographic origins") +
    theme(plot.title = element_text(hjust = 0.05))
```

Now, combining both plots.

```{r combine_plots}
final_pav <- ggpubr::ggarrange(pav_plot, hm_ggplot, nrow=2, 
                               labels = c("A", "B"))

####
# Saving and loading in different machines
save(final_pav, file = here("products", "plots", "pav_figure.rda"),
     compress="xz")
load(here("products", "plots", "pav_figure.rda"))
####

ggsave(final_pav,
       file=here::here("products", "plots", "PAV_analysis.png"),
       dpi=600, width=14, height=9)
```

## Session information

This document was created under the following conditions:

```{r session_info}
sessionInfo()
```
